backend/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# WebSocket
websockets==12.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Rate Limiting
slowapi==0.1.9

# External Services
twilio==8.11.1
msal==1.26.0
pyjwt==2.8.0
deepgram-sdk==3.2.0
google-generativeai==0.3.2
pinecone-client==3.0.2
elevenlabs==0.2.27

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Utilities
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

